Domain implementations

+++ src/fuzzy/domain/Domain.h
@@ -13,7 +13,7 @@ using namespace std;
 class Domain : public enable_shared_from_this<Domain>{
 public:
 	virtual uint getCardinality() = 0;
-    virtual uint getNumberOfComponents() = 0;
+    virtual unsigned long getNumberOfComponents() = 0;
 	virtual shared_ptr<Domain> getComponent(uint index) = 0;
     virtual DomainElement getElementAt(uint index) = 0;
     virtual uint indexOfElement (DomainElement element) = 0;
